Extract trace writer externalization helpers
